Updating work (fixing empty index.html, fixing docker, reenabling the the reuse of pre-downloaded repos)

+++ server/mysql/setup_battery_for_webserver.py
@@ -11,8 +11,9 @@ import tempfile
 
 # Get CLI parameters
 parser = argparse.ArgumentParser()
-parser.add_argument('--output', dest="output", help='Battery output directory',default=None)
+parser.add_argument('--output', dest="output", help='Battery output directory.')
 parser.add_argument('--experiments', dest="experiments", help='Experiment(s) to use in the battery, separated by commas.')
+parser.add_argument('--workingdir', dest="workingdir", help='Working directory, existing GIT repos are used if present and are not removed after execution.',default=None)
 
 try:
     args = parser.parse_args()
@@ -20,31 +21,45 @@ except:
     parser.print_help()
     sys.exit(0)
 
-if args.output != None:
-    if os.path.exists(args.output):
-        print("Output directory should not exist, remove and run again.")
-        sys.exit(0)
+if args.output == None or args.experiments == None:
+    print("Output and experiments parameters are mandatory.")
+    parser.print_help()
+    sys.exit(1)
 
+if os.path.exists(args.output):
+    print("Output directory should not exist, remove and run again.")
+    sys.exit(1)
 
-# We will set up the repo downloads, etc, in a temporary place
-working_dir = tempfile.mkdtemp()
-experiments = ",".split(args.experiments)
+battery_dest = args.output
+working_dir = args.workingdir
+is_temp_working_dir = False
 
-# Download the missing repos
+script_dirname=os.path.dirname(__file__)
+if script_dirname == "":
+    script_dirname="."
+
+if working_dir != None:
+    if not (os.path.isdir(working_dir)):
+        print("Working directory does not exist!")
+        sys.exit(1)
+else:
+    is_temp_working_dir = True
+    working_dir = tempfile.mkdtemp() # We will set up the repo downloads, etc, in a temporary place
+
+experiments = args.experiments.split(",")
+
+# Download the missing repos to working directory
 battery_repo='battery'
 experiment_repo='experiments'
 survey_repo='surveys'
 game_repo='games'
 
-# Download repos to working directory
 print('Downloading expfactory repos...')
 for repo in (battery_repo, experiment_repo, survey_repo, game_repo):
     folder = "%s/%s" %(working_dir,repo)
     if not (os.path.isdir(folder)):
-        download_repo(repo_type=repo, 
+        download_repo(repo_type=repo,
                       destination=folder)
-
-battery_dest = "%s/%s" %(working_dir,battery_repo)
 
 # Generate battery with the experiment(s)
 print('Generating base...')
@@ -56,17 +71,18 @@ base = generate_base(battery_dest=battery_dest,
                      add_experiments=True,
                      add_surveys=False,
                      add_games=False,
-                     battery_repo=battery_dest)
+                     battery_repo="%s/%s" %(working_dir,battery_repo))
 
 # Customize the battery
 custom_variables = dict()
 custom_variables["load"] = [("[SUB_TOTALTIME_SUB]", 30)]
 
-template_exp = "templates/webserver-battery-template.html"
-template_exp_output = "%s/index.html" %(base["battery_repo"])
+template_exp = ("%s/templates/webserver-battery-template.html" %(script_dirname))
+template_exp_output = "%s/index.html" %(battery_dest)
 
 print('Generating experiment and battery templates...')
-template_experiments(battery_dest=base["battery_repo"],
+
+template_experiments(battery_dest=battery_dest,
                      battery_repo=base["battery_repo"],
                      valid_experiments=base["experiments"],
                      custom_variables=custom_variables,
@@ -74,15 +90,13 @@ template_experiments(battery_dest=base["battery_repo"],
                      template_exp_output=template_exp_output)
 
 # Copy needed php files there
-add_files = glob("%s/*.php" %os.getcwd())
+add_files = glob("%s/*.php" %(script_dirname))
 for add_file in add_files:
-    shutil.copyfile(add_file,"%s/%s" %(base['battery_repo'],
-                                       os.path.basename(add_file)))
+    shutil.copy(add_file, battery_dest)
 
-# Remove git data from the battery
-if args.output != None:
-    copy_directory(base["battery_repo"],args.output)    
-    battery_dest = args.output
+# Cleanup
+shutil.rmtree("%s/.git" %(battery_dest)) # Remove git data from the battery
+if is_temp_working_dir:
+    shutil.rmtree(working_dir)
 
 print('Battery generated in %s' %(battery_dest))
-shutil.rmtree(working_dir)
